refactor(desktop): change hyper-bindings to a 'multi-module'
Problem statement:

Previous (3328d7d) change actually relied on xremap having a Hyper
binding emitted whenever I held the CapsLock physical button.

Xremap is configured in a NixOS system module as it's used independently
of hyprland too.

Solution:

'multi-module' (a term that does not exist and one I made up). I created
`app-jumps.nix` file that is actually a simple attrset with two
module-producing functions:

- `nixosModule`
- `homeManagerModule`

It is located on the top level beside `nixosModules` and
`homeManagerModules` and imported by both.

One downside of this change (reflected in a comment in `xremap` module)
is the way CapsLock escape remap also has to be here.